增加存在None值字段时抽象语法树节点的等于和哈希值的测试用例
